ROC-2161: moved ClaimAmountBreakdown and ClaimAmountRow classes to features/claim/form/models
