+++ src/main/app/utils/interestUtils.ts
@@ -1,7 +1,7 @@
 import InterestDateType from 'app/common/interestDateType'
 import { calculateInterest } from 'app/common/calculateInterest'
 import { MomentFactory } from 'common/momentFactory'
-import { ClaimAmountBreakdown } from 'app/forms/models/claimAmountBreakdown'
+import { ClaimAmountBreakdown } from 'features/claim/form/models/claimAmountBreakdown'
 import DraftClaim from 'app/drafts/models/draftClaim'
 
 export function interestAmount (claimDraft: DraftClaim): number {
