+++ src/main/features/claim/form/models/claimAmountRow.ts
@@ -1,9 +1,9 @@
 import * as _ from 'lodash'
 import { IsDefined, Min, ValidateIf } from 'class-validator'
 
-import { IsNotBlank } from 'forms/validation/validators/isBlank'
-import { Fractions } from 'forms/validation/validators/fractions'
-import { MaxLength } from 'forms/validation/validators/maxLengthValidator'
+import { IsNotBlank } from 'app/forms/validation/validators/isBlank'
+import { Fractions } from 'app/forms/validation/validators/fractions'
+import { MaxLength } from 'app/forms/validation/validators/maxLengthValidator'
 
 export class ValidationErrors {
   static readonly REASON_REQUIRED: string = 'Enter a reason'
