+++ src/main/features/claim/routes/amount.ts
@@ -5,7 +5,7 @@ import { Paths } from 'claim/paths'
 import { Form } from 'forms/form'
 import { FormValidator } from 'forms/validation/formValidator'
 import { ClaimValidator } from 'app/utils/claimValidator'
-import { ClaimAmountBreakdown, MAX_NUMBER_OF_EVENTS } from 'forms/models/claimAmountBreakdown'
+import { ClaimAmountBreakdown, MAX_NUMBER_OF_ROWS } from 'claim/form/models/claimAmountBreakdown'
 
 import { ErrorHandling } from 'common/errorHandling'
 import { DraftService } from 'common/draft/draftService'
@@ -14,7 +14,7 @@ function renderView (form: Form<ClaimAmountBreakdown>, res: express.Response): v
   res.render(Paths.amountPage.associatedView, {
     form: form,
     totalAmount: form.model.totalAmount(),
-    canAddMoreRows: form.model.rows.length < MAX_NUMBER_OF_EVENTS
+    canAddMoreRows: form.model.rows.length < MAX_NUMBER_OF_ROWS
   })
 }
