+++ src/test/app/pdf/mappers/interestMapper.ts
@@ -6,7 +6,7 @@ import Claim from 'app/claims/models/claim'
 import { MomentFactory } from 'common/momentFactory'
 import { InterestMapper } from 'app/pdf/mappers/interestMapper'
 import { MomentFormatter } from 'app/utils/momentFormatter'
-import { ClaimAmountBreakdown } from 'forms/models/claimAmountBreakdown'
+import { ClaimAmountBreakdown } from 'claim/form/models/claimAmountBreakdown'
 
 const creationDate = MomentFactory.currentDateTime()
 const customInterestDate = creationDate.subtract(10, 'days')
