+++ src/test/http-mocks/draft-store.ts
@@ -16,8 +16,8 @@ import Payment from 'app/pay/payment'
 import { Address } from 'forms/models/address'
 import DateOfBirth from 'app/forms/models/dateOfBirth'
 import { LocalDate } from 'forms/models/localDate'
-import { ClaimAmountBreakdown } from 'forms/models/claimAmountBreakdown'
-import { ClaimAmountRow } from 'forms/models/claimAmountRow'
+import { ClaimAmountBreakdown } from 'claim/form/models/claimAmountBreakdown'
+import { ClaimAmountRow } from 'claim/form/models/claimAmountRow'
 import Interest from 'forms/models/interest'
 import InterestDate from 'forms/models/interestDate'
 import Reason from 'forms/models/reason'
